moved quickstart into samples folder for testing
